move emoji getter for better organization
